move addPlayer from instance method to util
