+++ utils.js
@@ -1,3 +1,5 @@
+const { sortBy } = require("lodash");
+
 const getRandomInt = (min, max) => {
   min = Math.ceil(min);
   max = Math.floor(max);
@@ -15,6 +17,29 @@ const createGameId = () => {
     .join("");
 };
 
+const addPlayerToTeam = ({ teams, name }) => {
+  const teamWithFewerPlayers = sortBy(teams, team => {
+    return team.players.length;
+  })[0];
+
+  const assignedTeam = teamWithFewerPlayers;
+  const player = {
+    teamId: assignedTeam.id,
+    name
+  };
+  const newTeam = {
+    ...assignedTeam,
+    players: [...assignedTeam.players, player]
+  };
+  return teams.map(team => {
+    if (team.id === assignedTeam.id) {
+      return newTeam;
+    }
+    return team;
+  });
+};
+
 module.exports = {
-  createGameId
+  createGameId,
+  addPlayerToTeam
 };
